refactor: made `finite_differences` a private utility module for now
